Refactor SchemArrow to "PandasArrowConverter"
The main class name has been changed from "SchemArrow" to "PandasArrowConverter" for clarity. Also, dependency version for 'hypothesis' has been updated to "6.99.1". Paths in ci.yml and package information in pyproject.toml file have also been updated accordingly.
