+++ .github/workflows/ci.yml
@@ -2,7 +2,7 @@ name: CI
 on:
   pull_request:
     paths:
-      - 'schemarrow/**'
+      - 'pandas_pyarrow/**'
       - 'tests/**'
       - '.github/workflows/ci.yml'
   workflow_dispatch:
@@ -36,7 +36,7 @@ jobs:
           python-version: '3.11'
           cache: poetry
       - run: poetry install
-      - run: poetry run pytest --cov=schemarrow --cov-report=xml
+      - run: poetry run pytest --cov=pandas_pyarrow --cov-report=xml
       - uses: codecov/codecov-action@v2
         with:
           token: ${{ secrets.CODECOV_TOKEN }}
