--- schemarrow/__init__.py
@@ -1,3 +1,0 @@
-from .schema_arrow import SchemArrow
-
-__all__ = ["SchemArrow"]
