+++ tests/intg/test_db_date_dtypes.py
@@ -1,4 +1,4 @@
-from schemarrow.schema_arrow import SchemArrow
+from pandas_pyarrow.pda_converter import PandasArrowConverter
 
 import db_dtypes as dbdt
 import pandas as pd
@@ -20,7 +20,7 @@ from parametrization import Parametrization
     expected_dtype="time64[us][pyarrow]",
 )
 def test_db_date_dtypes(df_data, expected_dtype):
-    sa = SchemArrow()
+    sa = PandasArrowConverter()
     adf = sa(df_data)
 
     assert list(adf.dtypes)[0] == expected_dtype
