+++ tests/unit/property_based/test_numeric.py
@@ -1,7 +1,7 @@
 from typing import Tuple
 
-from schemarrow import SchemArrow
-from schemarrow.mappers import numeric_mapper
+from pandas_pyarrow import PandasArrowConverter
+from pandas_pyarrow.mappers import numeric_mapper
 from tests.unit.property_based.pb_sts import single_column_df_st
 
 import hypothesis as hp
@@ -19,7 +19,7 @@ import pandas as pd
     )
 )
 def test_float_numpy_api_hp(pair: Tuple[pd.DataFrame, str]):
-    sa = SchemArrow()
+    sa = PandasArrowConverter()
     df, target_dtype = pair
     adf = sa(df)
 
@@ -37,7 +37,7 @@ def test_float_numpy_api_hp(pair: Tuple[pd.DataFrame, str]):
     )
 )
 def test_float_array_api_hp(pair: Tuple[pd.DataFrame, str]):
-    sa = SchemArrow()
+    sa = PandasArrowConverter()
     df, target_dtype = pair
     adf = sa(df)
 
@@ -55,7 +55,7 @@ def test_float_array_api_hp(pair: Tuple[pd.DataFrame, str]):
     )
 )
 def test_int_numpy_api_hp(pair: Tuple[pd.DataFrame, str]):
-    sa = SchemArrow()
+    sa = PandasArrowConverter()
     df, target_dtype = pair
     adf = sa(df)
 
@@ -73,7 +73,7 @@ def test_int_numpy_api_hp(pair: Tuple[pd.DataFrame, str]):
     )
 )
 def test_int_array_api_hp(pair: Tuple[pd.DataFrame, str]):
-    sa = SchemArrow()
+    sa = PandasArrowConverter()
     df, target_dtype = pair
     adf = sa(df)
